pyproject.toml
[tool.poetry]
name = "unisite"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Ruslan Bagymanov <ruslanbagymanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
asyncpg = "^0.29.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
jwt = "^1.3.1"
datetime = "^5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
